+++ wallet/commands/commands.go
@@ -95,6 +95,8 @@ func CheckSubmitTransactionRequest(req *walletpb.SubmitTransactionRequest) comma
 		cmdErr = commands.CheckApplyReferralCode(cmd.ApplyReferralCode)
 	case *walletpb.SubmitTransactionRequest_UpdateMarginMode:
 		cmdErr = commands.CheckUpdateMarginMode(cmd.UpdateMarginMode)
+	case *walletpb.SubmitTransactionRequest_JoinTeam:
+		cmdErr = commands.CheckJoinTeam(cmd.JoinTeam)
 	default:
 		errs.AddForProperty("input_data.command", commands.ErrIsNotSupported)
 	}
@@ -233,6 +235,10 @@ func WrapRequestCommandIntoInputData(data *commandspb.InputData, req *walletpb.S
 	case *walletpb.SubmitTransactionRequest_UpdateMarginMode:
 		data.Command = &commandspb.InputData_UpdateMarginMode{
 			UpdateMarginMode: req.GetUpdateMarginMode(),
+		}
+	case *walletpb.SubmitTransactionRequest_JoinTeam:
+		data.Command = &commandspb.InputData_JoinTeam{
+			JoinTeam: req.GetJoinTeam(),
 		}
 	default:
 		panic(fmt.Sprintf("command %T is not supported", cmd))